2-Gleb/train/requirements.txt
pillow == 8.0.1
pandas == 1.1.5
shapely == 1.7.1
rasterio == 1.2.0
requests == 2.25.0
numpy == 1.19.4
opencv-python == 4.4.0.46
albumentations == 0.5.2
fastprogress == 1.0.0
tqdm == 4.56.0
matplotlib == 3.3.3
torch == 1.7.0
torchvision == 0.8.1
torchsummary == 1.5.1
scipy == 1.5.4
yacs == 0.1.8
loguru == 0.5.3
